+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add support for dumping a database using the CLI (use the `--dump` flag)
+  ([#88](https://github.com/go-testfixtures/testfixtures/pull/88), [#63](https://github.com/go-testfixtures/testfixtures/issues/63)).
 - Support SkipResetSequences and ResetSequencesTo for MySQL and MariaDB
   ([#91](https://github.com/go-testfixtures/testfixtures/pull/91)).